introduced version with staged functions
Precompile didn't work with @aaalexandrov's version. I took this as a
motivation to make the staged function approach work, as it was quick to
do and actually solved the precompilation problem. We could just switch
out the macro depending on the julia version. Besides regressions in
stack traces (which hopefully get solved), I don't know anything that
speak against this approach.